Addressed bug when not set right amount and dupe items.
Fixed a bug where the item amount was not set correctly. When two or more players are inside the same inventory and they add more than one item while the shallCheckDuplicates setting is true, it creates duplicate items. When one player leaves, they receive the correct amount of items, but the extra items are not cleared from the inventory.