Added some improvements when serializing data
Removed the deprecated annotation from the old methods that not require you to set the file.

Now, you are no longer required to use the method that involves providing the file argument, as it will be handled internally. However, you can still use the method where you can provide the file argument if you want to save to own specified file.