Improve security and make it more easy to use.
Clean up the code to make it more easy to understand how to use it and removed need for self tell what type of database you use.

Improve security when you save data and should offer little better performance when you save large batch.

It will also try to preserve the value you want to save and not convert everything to string.

Fixed error when PostgreSQL could not create the database and fixed error when database could not find the table.