Improve the NBT
Should give better performance, it will however break some combability if you really on the old getMetadata() method where you can set the class to cast to.

You can use the new  getMetadata(ItemStack) method and use one of the methods in NBTValueWrapper class, but it does not support so many classes.