Fixed the nbt and inventory title.
Fixed inventory title update for 1.20.2. Will be more changes, because spigot added official support for update the title and hopefully I don't need to update this for future versions. This is not jet implemented and need to test it out so it does not break the inventory when use it asynchronous.

And fixed the problem when not set the metadata.