Improve the consistency when use the plugin
Change how the menus for the utility plugin works, you now only need to sett the plugin instance in the constructor not where you chose to open the inventory.

Also make the Javadoc clearer how the normal MenuHolder class works and it does not use the list of items any more you have to implement MenuHolderPage. The MenuHolder contains the option to implement your own paged menu, but you have to do more manually work. Check the wiki for latest info.