Clean up the commands
The old CommandBuilder class was only confusing to use, replace with CommandProperty and you register your CommandHolder directly.

The CommandHandler will also be removed due to questionable usefulness in the code and also the CommandRegistering interface is questionable also.

Made some bigger improvements on the javadocs.

Will update the wiki soon when everything is tested.