Fix so it set marker placeholder global
Makes it safer by preventing a mix of placeholder-based and value-based SQL parts, which could cause errors when executing the query. This change reduces the chance of issues when using either prepared statements or raw queries with values directly embedded. In most cases, mixing is now disallowed.

You can still bypass this restriction by calling the constructor directly, but that is not recommended.